+++ blocks/carousel/carousel.js
@@ -403,15 +403,3 @@ export default async function createCarousel(block, data, config) {
   await carousel.render();
   return carousel;
 }
-
-export function summariseDescription(description, charCount) {
-  let result = description;
-  if (result.length > charCount) {
-    result = result.substring(0, charCount);
-    const lastSpaceIndex = result.lastIndexOf(' ');
-    if (lastSpaceIndex !== -1) {
-      result = result.substring(0, lastSpaceIndex);
-    }
-  }
-  return `${result}…`;
-}
