+++ scripts/scripts.js
@@ -41,6 +41,24 @@ export function loadScript(url, callback, type, async) {
 }
 
 /**
+ * Summarises the description to maximum character count without cutting words.
+ * @param {string} description Description to be summarised
+ * @param {number} charCount Max character count
+ * @returns summarised string
+ */
+export function summariseDescription(description, charCount) {
+  let result = description;
+  if (result.length > charCount) {
+    result = result.substring(0, charCount);
+    const lastSpaceIndex = result.lastIndexOf(' ');
+    if (lastSpaceIndex !== -1) {
+      result = result.substring(0, lastSpaceIndex);
+    }
+  }
+  return `${result}…`;
+}
+
+/**
  * Will add the 'text-caption' class to the empasised elements from
  * the next sibling to mark them as captions
  * @param {NodeListOf<Element>} elems Elements which are presumed to have a caption attached.
