+++ templates/blog/blog.js
@@ -1,4 +1,4 @@
-import { styleCaption } from '../../scripts/scripts.js';
+import { styleCaption, summariseDescription } from '../../scripts/scripts.js';
 import {
   a, div, h3, p,
 } from '../../scripts/dom-helpers.js';
@@ -50,18 +50,6 @@ export default function buildAutoBlocks() {
   );
 }
 
-function summariseDescription(description) {
-  let result = description;
-  if (result.length > 75) {
-    result = result.substring(0, 75);
-    const lastSpaceIndex = result.lastIndexOf(' ');
-    if (lastSpaceIndex !== -1) {
-      result = result.substring(0, lastSpaceIndex);
-    }
-  }
-  return `${result}…'`;
-}
-
 export function renderBlogCard(item) {
   const itemImage = item.thumbnail && item.thumbnail !== '0' ? item.thumbnail : item.image;
   const buttonText = item.cardC2A && item.cardC2A !== '0' ? item.cardC2A : 'Read More';
@@ -77,7 +65,7 @@ export function renderBlogCard(item) {
         h3(
           a({ href: item.path }, item.title),
         ),
-        p({ class: 'blog-card-description' }, summariseDescription(item.description)),
+        p({ class: 'blog-card-description' }, summariseDescription(item.description, 75)),
         p({ class: 'button-container' },
           a({ href: item.path, 'aria-label': buttonText, class: 'button primary' }, buttonText),
         ),
